app/requirements.txt
SQLAlchemy==2.0.25
fastapi==0.108.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
uvicorn==0.25.0